+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Table of Contents
 
+- [1.2.14 - 2025-12-19](#1214---2025-12-19)
 - [1.2.13 - 2025-12-05](#1213---2025-12-05)
 - [1.2.12 - 2025-11-12](#1212---2025-11-12)
 - [1.2.11 - 2025-10-27](#1211---2025-10-27)
@@ -47,6 +48,18 @@ All notable changes to this project will be documented in this file. The format 
 - [1.1.1 - 2024-08-28](#111---2024-08-28)
 - [1.1.0 - 2024-08-19](#110---2024-08-19)
 - [1.0.0 - 2024-06-06](#100---2024-06-06)
+
+## [1.2.14] - 2025-12-19
+
+### Added
+- `MerklePath.FindLeafByOffset()` method to find a PathElement at a given offset and level
+- `MerklePath.AddLeaf()` method to add a PathElement to a specified level
+- `MerklePath.ComputeMissingHashes()` method to compute parent hashes from sparse merkle path data
+
+### Changed
+- `Outpoint.Bytes()` now returns 36 bytes in little-endian format (consistent with transaction format)
+- `Outpoint.TxBytes()` is now an alias for `Bytes()` for backward compatibility
+- `NewOutpointFromBytes()` now accepts a `[]byte` slice instead of `[36]byte` array
 
 ## [1.2.13] - 2025-12-05